try to refactor location picker :(
